changing delete function

+++ js/home-controller.js
@@ -47,15 +47,19 @@ myApp.config(function($routeProvider) {
       // }
 
       $scope.remove = function (index){
+        if (window.confirm("Do you really want to delete?")) { 
+          
+          Newconcerts = JSON.parse(window.localStorage.getItem('concerts'));
+          Newchart = JSON.parse(window.localStorage.getItem('chartItem'));
+          Newconcerts.splice(index,1);
+          Newchart.splice(index,1);
+          window.localStorage.setItem('concerts', JSON.stringify(Newconcerts));
+          window.localStorage.setItem('chartItem', JSON.stringify(Newchart));
+          $scope.Newconcerts = JSON.parse(window.localStorage.getItem('concerts'));
+          loadChartHOme();
+          
+        }
         
-        Newconcerts = JSON.parse(window.localStorage.getItem('concerts'));
-        Newchart = JSON.parse(window.localStorage.getItem('chartItem'));
-        Newconcerts.splice(index,1);
-        Newchart.splice(index,1);
-        window.localStorage.setItem('concerts', JSON.stringify(Newconcerts));
-        window.localStorage.setItem('chartItem', JSON.stringify(Newchart));
-        $scope.Newconcerts = JSON.parse(window.localStorage.getItem('concerts'));
-        loadChartHOme();
       }
